Integrate unit tests to docker-compose and py.test fixtures
With this commit, the fixtures of py.test that provide access to the app and the
db will work correctly. Moreover, I have refined the configuration so that unit
tests are now a service in the docker-compose structure.

Since there was a problem concerning the database models and a circular
dependency, I had to change a relationship so that it is saved with an ``ALTER``
instruction to break the circular dependency. I have decided to group all data
API models in a single migration since there it simplifies model modifications.