Add basic fixtures and some initial unit tests
This commit contains the modifications needed for celery to work correctly under
unit tests. That is, there is a tricky situation concerning the Flask app
context and the fact that celery tasks work under a new Flask app context. This
is usually managed by the Flask-celery-helper package, but there is a problem
when running celery tasks on eager mode (useful for unit tests, even if the
celery documentation says that it should not be used; I disagree).

This commit also creates some useful fixtures for pytest, which are the Flask
application, the SQLAlchemy instance object and a database session. These last
two should now clean the database after the unit tests finish. Unit tests that
use the database session should add the `session` fixture so that a session is
created and cleaned. Otherwise, there `db` fixture will block during the
teardown process because of dangling sessions.