Change the file create specification to a simple multipart
See the notes.rst on some conclusions leading to this change. Briefly, we cannot
have a file create endpoint that accepts the file contents *and* the file
metadata *at the same time*. This is regrettable, because it is the exact
use-case of a multipart endpoint. There are some solutions, all very
complicated. For this reason, we will only accept the file contents for the
moment and if we add the file metadata at the same endpoint in the future, we
could extend the current multipart definition.