Add implementation of workspace scan and commit
Workspace scan is implemented with a clean version of the old SQL code, now
using SQLAlchemy for the whole operation. This also depends heavily on the query
for the workspace metadata, which has been refactored into the workspace model.
However, this does not take into account the exact family versions; it is
currently only working with the latest version.

The commit implementation does not have a conflict detection yet.

The query implementation was started but due to a problem on the API design, it
is highly advised to change the API so that the query is not a GET.