Add persistent volume claims and general reoganization of k8s files
This commit updates the k8s files for a better metadata declaration, usage of
persistent volume claims, and update of the documentation to reflect these new
changes.

By using persistent volume claims, the installation procedure is simpler. But in
order to keep the disks when they are not used (and not be reclaimed and
erased), a specific storage class element needs to be declared as well.