Add file delete soft implementation
This commit adds a new metadata entry for the base family: state. It can be
READY (file is usable), TEMPORARY (file is usable but will not be saved to the
global workspace if a commit occurs) and DELETE (file will be deleted from
global workspace if it existed).

A new endpoint DELETE file is also implemented, which changes the metadata but
does not do any bucket operation yet.